Devie functionality + Load Commands on start
This commit adds the functionality from the `devie-bot` to the `resource-bot` in preparation to merge the two projects.
This includes adding a new event handler `messageReactionAdd.ts` and associated environment settings.
The new event handler requires access to some `partials` as reactions can be added to old messages which will need to be fetched.
`partials: ['MESSAGE', 'CHANNEL', 'REACTION'],`
These were copied directly from the current `devie-bot` implementation.
It also requires a new *intent* so that it can access _reactions_ added to _guild_messages_.
`Intents.FLAGS.GUILD_MESSAGE_REACTIONS,`

Additionally this commit makes the bot register all commands on startup, so that there is not a separate process for updating the command list.
Loading the commands has been moved into the discord client and will be executed on startup.
The index file was upated to allow the async function to be `await`ed.